Refactor Utilities

+++ Assets/Downloads/Devion Games/Utilities/Scripts/Editor/Property Drawers/MinMaxSliderDrawer.cs
@@ -39,7 +39,7 @@ namespace DevionGames
 
 		public static float Round (float value, int digits)
 		{
-			float mult = Mathf.Pow (10.0f, (float)digits);
+			float mult = Mathf.Pow (10.0f, digits);
 			return Mathf.Round (value * mult) / mult;
 		}
